Rename ROBOCORP_BROWSERS to RPA_SELENIUM_BROWSER_ORDER, move browser_preferences()
